refactor: separate WindData and DataSource classes (#175)
