+++ prereise/cli/data_sources/__init__.py
@@ -1,0 +1,9 @@
+from prereise.cli.data_sources.wind_data import WindData
+
+
+def get_data_sources_list():
+    """Provides list of all data sources
+
+    :return: (*list*)
+    """
+    return [WindData()]
