+++ prereise/cli/data_sources/tests/test_data_source.py
@@ -1,0 +1,7 @@
+from prereise.cli.data_sources import get_data_sources_list
+from prereise.cli.data_sources.wind_data import WindData
+
+
+def test_get_data_sources_list():
+    data_sources_list = get_data_sources_list()
+    assert isinstance(data_sources_list[0], WindData)
